Fix FFT calculation in the sound modules
FFTModule is using realTransform, so it is already computing the
absolutefft with regards to the input. The second half of the output
array was being left completely zeroed, yet the absolute FFT module
was operating under the assumption that it was dealing with complex
number pairs.

Since the complex FFT is not used at all in the programs (and
probably has little utility in general), I removed the absolute
FFT module and made the FFT module set the absolutefft field in
the channel object.