Save presets, "program" type params and default-preset-override mechanism
Divide programConfig in defaults, presetOverrides and override.
Update UI to show when a parameters is default, from preset or override.
Add save preset functionality, which saves it in a json file
Add new type "program" to parameters, which allows to play in the ui with
the parameters of subprograms which are parameters of other programs.
Add new Mix program which makes a simple use case of this.